Move index_thread_qty to KNNCodecService
